[Security][Guard] Remove GuardAuthenticatorInterface
