+++ src/Symfony/Component/Security/CHANGELOG.md
@@ -11,6 +11,8 @@ CHANGELOG
  * removed support for voters that don't implement the `VoterInterface`
  * added a sixth `string $context` argument to `LogoutUrlGenerator::registerListener()`
  * removed HTTP digest authentication
+ * removed `GuardAuthenticatorInterface` in favor of `AuthenticatorInterface`
+ * removed `AbstractGuardAuthenticator::supports()`
 
 3.4.0
 -----
