+++ src/Symfony/Component/Security/Guard/AbstractGuardAuthenticator.php
@@ -11,7 +11,6 @@
 
 namespace Symfony\Component\Security\Guard;
 
-use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Component\Security\Guard\Token\PostAuthenticationGuardToken;
 
@@ -22,18 +21,6 @@ use Symfony\Component\Security\Guard\Token\PostAuthenticationGuardToken;
  */
 abstract class AbstractGuardAuthenticator implements AuthenticatorInterface
 {
-    /**
-     * {@inheritdoc}
-     *
-     * @deprecated since version 3.4, to be removed in 4.0
-     */
-    public function supports(Request $request)
-    {
-        @trigger_error(sprintf('The "%s()" method is deprecated since version 3.4 and will be removed in 4.0. Implement the "%s::supports()" method in class "%s" instead.', __METHOD__, AuthenticatorInterface::class, get_class($this)), E_USER_DEPRECATED);
-
-        return true;
-    }
-
     /**
      * Shortcut to create a PostAuthenticationGuardToken for you, if you don't really
      * care about which authenticated token you're using.
