+++ src/Symfony/Component/Security/Guard/Firewall/GuardAuthenticationListener.php
@@ -15,10 +15,9 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
 use Symfony\Component\Security\Guard\GuardAuthenticatorHandler;
-use Symfony\Component\Security\Guard\GuardAuthenticatorInterface;
+use Symfony\Component\Security\Guard\AuthenticatorInterface;
 use Symfony\Component\Security\Guard\Token\PreAuthenticationGuardToken;
 use Symfony\Component\Security\Core\Authentication\AuthenticationManagerInterface;
-use Symfony\Component\Security\Guard\AuthenticatorInterface;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Core\Exception\AuthenticationException;
@@ -94,7 +93,7 @@ class GuardAuthenticationListener implements ListenerInterface
         }
     }
 
-    private function executeGuardAuthenticator($uniqueGuardKey, GuardAuthenticatorInterface $guardAuthenticator, GetResponseEvent $event)
+    private function executeGuardAuthenticator($uniqueGuardKey, AuthenticatorInterface $guardAuthenticator, GetResponseEvent $event)
     {
         $request = $event->getRequest();
         try {
@@ -103,27 +102,14 @@ class GuardAuthenticationListener implements ListenerInterface
             }
 
             // abort the execution of the authenticator if it doesn't support the request
-            if ($guardAuthenticator instanceof AuthenticatorInterface) {
-                if (!$guardAuthenticator->supports($request)) {
-                    return;
-                }
-                // as there was a support for given request,
-                // authenticator is expected to give not-null credentials.
-                $credentialsCanBeNull = false;
-            } else {
-                // deprecated since version 3.4, to be removed in 4.0
-                $credentialsCanBeNull = true;
+            if (!$guardAuthenticator->supports($request)) {
+                return;
             }
 
             // allow the authenticator to fetch authentication info from the request
             $credentials = $guardAuthenticator->getCredentials($request);
 
             if (null === $credentials) {
-                // deprecated since version 3.4, to be removed in 4.0
-                if ($credentialsCanBeNull) {
-                    return;
-                }
-
                 throw new \UnexpectedValueException(sprintf('The return value of "%s::getCredentials()" must not be null. Return false from "%s::supports()" instead.', get_class($guardAuthenticator), get_class($guardAuthenticator)));
             }
 
@@ -196,7 +182,7 @@ class GuardAuthenticationListener implements ListenerInterface
      * @param TokenInterface         $token
      * @param Response               $response
      */
-    private function triggerRememberMe(GuardAuthenticatorInterface $guardAuthenticator, Request $request, TokenInterface $token, Response $response = null)
+    private function triggerRememberMe(AuthenticatorInterface $guardAuthenticator, Request $request, TokenInterface $token, Response $response = null)
     {
         if (null === $this->rememberMeServices) {
             if (null !== $this->logger) {
