+++ src/Symfony/Component/Security/Guard/Tests/Provider/GuardAuthenticationProviderTest.php
@@ -87,75 +87,13 @@ class GuardAuthenticationProviderTest extends TestCase
     }
 
     /**
-     * @group legacy
-     */
-    public function testLegacyAuthenticate()
-    {
-        $providerKey = 'my_cool_firewall';
-
-        $authenticatorA = $this->getMockBuilder('Symfony\Component\Security\Guard\GuardAuthenticatorInterface')->getMock();
-        $authenticatorB = $this->getMockBuilder('Symfony\Component\Security\Guard\GuardAuthenticatorInterface')->getMock();
-        $authenticatorC = $this->getMockBuilder('Symfony\Component\Security\Guard\GuardAuthenticatorInterface')->getMock();
-        $authenticators = array($authenticatorA, $authenticatorB, $authenticatorC);
-
-        // called 2 times - for authenticator A and B (stops on B because of match)
-        $this->preAuthenticationToken->expects($this->exactly(2))
-            ->method('getGuardProviderKey')
-            // it will return the "1" index, which will match authenticatorB
-            ->will($this->returnValue('my_cool_firewall_1'));
-
-        $enteredCredentials = array(
-            'username' => '_weaverryan_test_user',
-            'password' => 'guard_auth_ftw',
-        );
-        $this->preAuthenticationToken->expects($this->atLeastOnce())
-            ->method('getCredentials')
-            ->will($this->returnValue($enteredCredentials));
-
-        // authenticators A and C are never called
-        $authenticatorA->expects($this->never())
-            ->method('getUser');
-        $authenticatorC->expects($this->never())
-            ->method('getUser');
-
-        $mockedUser = $this->getMockBuilder('Symfony\Component\Security\Core\User\UserInterface')->getMock();
-        $authenticatorB->expects($this->once())
-            ->method('getUser')
-            ->with($enteredCredentials, $this->userProvider)
-            ->will($this->returnValue($mockedUser));
-        // checkCredentials is called
-        $authenticatorB->expects($this->once())
-            ->method('checkCredentials')
-            ->with($enteredCredentials, $mockedUser)
-            // authentication works!
-            ->will($this->returnValue(true));
-        $authedToken = $this->getMockBuilder('Symfony\Component\Security\Core\Authentication\Token\TokenInterface')->getMock();
-        $authenticatorB->expects($this->once())
-            ->method('createAuthenticatedToken')
-            ->with($mockedUser, $providerKey)
-            ->will($this->returnValue($authedToken));
-
-        // user checker should be called
-        $this->userChecker->expects($this->once())
-            ->method('checkPreAuth')
-            ->with($mockedUser);
-        $this->userChecker->expects($this->once())
-            ->method('checkPostAuth')
-            ->with($mockedUser);
-
-        $provider = new GuardAuthenticationProvider($authenticators, $this->userProvider, $providerKey, $this->userChecker);
-        $actualAuthedToken = $provider->authenticate($this->preAuthenticationToken);
-        $this->assertSame($authedToken, $actualAuthedToken);
-    }
-
-    /**
      * @expectedException \Symfony\Component\Security\Core\Exception\BadCredentialsException
      */
     public function testCheckCredentialsReturningNonTrueFailsAuthentication()
     {
         $providerKey = 'my_uncool_firewall';
 
-        $authenticator = $this->getMockBuilder('Symfony\Component\Security\Guard\GuardAuthenticatorInterface')->getMock();
+        $authenticator = $this->getMockBuilder(AuthenticatorInterface::class)->getMock();
 
         // make sure the authenticator is used
         $this->preAuthenticationToken->expects($this->any())
